SDK-642: Create "Attribute" folder, move classes
