+++ src/Yoti.Auth/GlobalSuppressions.cs
@@ -15,5 +15,5 @@
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Major Code Smell", "S3925:\"ISerializable\" should be implemented correctly", Justification = "[ISerializable] not available in 'netstandard' target framework", Scope = "type", Target = "~T:Yoti.Auth.Exceptions.AmlException")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Major Code Smell", "S3925:\"ISerializable\" should be implemented correctly", Justification = "[ISerializable] not available in 'netstandard' target framework", Scope = "type", Target = "~T:Yoti.Auth.Exceptions.YotiProfileException")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Major Code Smell", "S3925:\"ISerializable\" should be implemented correctly", Justification = "[ISerializable] not available in 'netstandard' target framework", Scope = "type", Target = "~T:Yoti.Auth.Exceptions.DynamicShareException")]
-[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "We don't want any exception to stop the remaining attributes from being parsed, so we need to catch the general exception", Scope = "member", Target = "~M:Yoti.Auth.AttributeConverter.ConvertToBaseAttributes(Yoti.Auth.ProtoBuf.Attribute.AttributeList)~System.Collections.Generic.Dictionary{System.String,Yoti.Auth.BaseAttribute}")]
+[assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "We don't want any exception to stop the remaining attributes from being parsed, so we need to catch the general exception", Scope = "member", Target = "~M:Yoti.Auth.Attribute.AttributeConverter.ConvertToBaseAttributes(Yoti.Auth.ProtoBuf.Attribute.AttributeList)~System.Collections.Generic.Dictionary{System.String,Yoti.Auth.Attribute.BaseAttribute}")]
 [assembly: System.Diagnostics.CodeAnalysis.SuppressMessage("Build", "CA1724:The type name Extension conflicts in whole or in part with the namespace name 'Org.BouncyCastle.X509.Extension'. Change either name to eliminate the conflict.", Justification = "Since this is a class library, clients can fully qualify the name to avoid this ambiguity", Scope = "type", Target = "~T:Yoti.Auth.ShareUrl.Extensions.Extension`1")]
