+++ src/Yoti.Auth/Profile/BaseProfile.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Yoti.Auth.Attribute;
 
 namespace Yoti.Auth.Profile
 {
