+++ src/Yoti.Auth/Profile/YotiProfile.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using Newtonsoft.Json.Linq;
+using Yoti.Auth.Attribute;
 using Yoti.Auth.Document;
 using Yoti.Auth.Images;
 using Yoti.Auth.Verifications;
