+++ src/Yoti.Auth/Verifications/AgeVerification.cs
@@ -2,6 +2,7 @@
 using System.Globalization;
 using System.Text.RegularExpressions;
 using Yoti.Auth.Anchors;
+using Yoti.Auth.Attribute;
 
 namespace Yoti.Auth.Verifications
 {
