+++ test/Yoti.Auth.Tests/ActivityTests.cs
@@ -6,6 +6,7 @@ using System.Text;
 using Google.Protobuf;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json.Linq;
+using Yoti.Auth.Attribute;
 using Yoti.Auth.Document;
 using Yoti.Auth.Images;
 using Yoti.Auth.Profile;
